Pc 7550 aggregate firebase offer events (#208)
* (PC-7550): correct name error ans split file

* (PC-7550): aggregate firebase offer events
